+++ src/Lib/Sort/sort.ts
@@ -1,3 +1,0 @@
-export interface ISort {
-	descending?: boolean;
-}