+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Device screenshot `listLastScreenshotsByDevices` method to fetch list of last screenshots for devices. One last screenshot for each device
+
 ## [1.14.0] - 2022-08-03
 ### Added
 - Listing organizations by `companyUid` in filter